// web/lib/api/generated/model/warning.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { WarningDetails } from './warningDetails';

export interface Warning {
  code: string;
  message: string;
  details: WarningDetails;
}